chore: Fix issues from Dart 3.13 upgrade
Dart 3.13 added dot shorthand, which Flutter 3.47 uses extensively. But, because we're using older versions of `json_serializable` (which include older versions of `analyzer`), running json generation on files that include that doc shorthand causes the generation to crash.

Fix by only running the json_serializable on files that need it, and moving RUM enums out of a file that includes `flutter` libraries so that `ddrum_events.dart` can import them cleanly.

Thankfully, this doesn't affect consumers of the library as we ship pre-generated files, and they do not have to run the generation themselves.

Also fix datadog_dio build issues with Flutter 3.47.
