+++ packages/datadog_flutter_plugin/build.yaml
@@ -5,3 +5,8 @@ targets:
         - example/**
         - integration_test_app/**
         - e2e_test_app/**
+    builders:
+      json_serializable:
+        generate_for:
+          - lib/src/logs/ddlog_event.dart
+          - lib/src/rum/ddrum_events.dart
