+++ packages/datadog_flutter_plugin/lib/src/rum/rum.dart
@@ -3,6 +3,7 @@
 // Copyright 2019-Present Datadog, Inc.
 
 export 'ddrum.dart';
+export 'ddrum_enums.dart';
 export 'ddrum_events.dart';
 export 'navigation_observer.dart';
 export 'resource_headers_extractor.dart';
