+++ packages/datadog_flutter_plugin/test/rum/rum_user_action_detector_test.dart
@@ -3,8 +3,7 @@
 // Copyright 2019-Present Datadog, Inc.
 
 import 'package:datadog_common_test/datadog_common_test.dart';
-import 'package:datadog_flutter_plugin/src/rum/ddrum.dart';
-import 'package:datadog_flutter_plugin/src/rum/rum_user_action_detector.dart';
+import 'package:datadog_flutter_plugin/src/rum/rum.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mocktail/mocktail.dart';
